+++ spec/spec_helper.rb
@@ -17,6 +17,7 @@ RSpec.configure do |config|
   config.before(:suite) do
     FactoryBot.find_definitions
   end
+
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!