+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- additional verification for token:
+    - verify token age
+    - verify sso params
+    - website signature
+- support for migration
+- settings model for jwt and sso params
+
+### Changed
+- all SSO params have prefix `sso`
+
+### Removed
+- `refId` is not longer part of token
+
 ## [0.0.5] - 2020-12-15
 ### Changed
 - add `0x` to signature to be compatible with JS and not throw `signature missing v and recoveryParam` error